Update scan form batch to not rely on scan forms to calculate the shipment count
Due to the way the Express for Stamps API works regarding scan forms, the scan
forms in ShipWorks do not necessarily have any handle to the underlying shipments
when retrieving a batch from the database. If we continued to rely on the
ScanForm.ShipmentCount, there's the potential that the shipment count in the
batch would be innacurate if there were multiple scan forms in the batch since
the shipments are related to the batch (i.e. prior to this change each form
would indicate the shipments for the entire batch resulting in the batch
thinking that shipment count = [count of scan forms in the batch] x [total shipments]

- Adding a FK relation from the Endicia and Stamps shipment tables to the
  scan form batch table

- Removed the ShipmentCount property from the ScanForm class, and dropped the
  ShipmentCount column from the EndiciaScanForm and StampsScanForm tables, and
  updated the various pieces of the scan form API in ShipWorks accordingly
  (carrier accounts, repositories, etc.)

- Regenerated data model entity classes

- Introduce an interface for abstracting the LLBLGen dependency from the
  actual ScanFormBatch domain object that is used to determine the total
  number of shipments in the batch rather than relying on the scan form

- Installation and update scripts modified accordingly based on SQL Compare