For KnowledgeBaseEntry, added overrides to ApplyTo and ApplyFrom to take a list of ShipmentCustomsEntityItem ApplyFrom first calls ApplyFrom(adapters), then clears the current list of kb customs items, then iterates through each ShipmentCustomsItems passed in, and adds each to the kb customs items list. ApplyTo first calls ApplyTo(adapters), then if there are any KnowledgeBaseCustomItems for the kb entry, it clears the current ShipmentCustomsItems list, and then adds a new ShipmentCustomsItem for each KB customs item.
Updated ShipmentType.ApplyShipSense to determine if the shipment is domestic, and if it is, call ApplyTo(adapters).

If it is international, it calls ApplyTo(adapters, customsItems).  If the shipment.CustomsItems is populated, ApplyShipSense then updates the shipment customs properties.

Updated ShippingManager.LogToShipSenseKnowledgebase to call the appropriate ApplyFrom method based on domestic/international.