Open UPS account error handling, same as pickup change, and set NoScheduledPickup as default and warn of possible fees
Update UPS Setup Wizard for the case when UPS opens an account but fails to register in ShipWorks.
When unable to register an account, don't popup an error and go to final wizard page.
Show an error that says what happened on final wizard page
List NoScheduledPickup first and have that be selected.
Change radio buttons that state pickup is the same as billing to a checkbox
Warn of possible fees if selecting something other than NoScheduledPickup.