Ensure shipment is still editable after canceling counter rates dialog
When trying to create a label with a counter rate account, you are presented with a dialog to create an account.  If you canceled this dialog when using the shipping panel, the panel would remain disabled until you clicked away and came back.
This also exposed a bug where the shipment would get stuck in a state where ShipWorks always thought it was modified and display the warning multiple times