Ensure focused control is saved when saving to database
If a textbox in the shipping panel has focus when selecting a new order, that textbox does not lose focus. This means that if the text property has a binding that updates when the textbox loses focus, that value doesn't update the view model when saving.

This commit also adds tests to the shipping panel view model for saving to the database and pares down the database calls in teh insurance view model tests to speed them up a bit