Fix issue where new shipment wouldn't load in panel
The address validator sets the order of the shipment because it may have had its address changed. This didn't take into account whether the shipment already had its order loaded. Because of this, the order associated with the shipment and the order that was loaded by the shipment loader were not the same instance