There was a crash in the obfuscated build when the store address was blank and tried getting counter rates.
The counter rates foot note for invalid store was using XAML with constructor parameters.  For some reason the obfuscated version didn't like this.  I think it might be because the footnotes are DataTemplates in a ResourceDiction, but I don't know for sure.  Changed it to use the VisibleWhenEqualToParameterConverter and it now works OK.