Rating panel rates were not being updated when the shipping settings dlg was closed.
Added a pipeline for refreshing rates.
Needed to use the Reactive Switch operator so that when a new rates retrieving message comes in, we abandon the sequence we were using and start over with the new one.  This way we don't keep around previously received rates retrieved messages and close the spinner prematurely.