Rates from the previously selected carrier were still being shown when switching to Other or None.
The RatesRetrieverService check that the changed field is a rating field.  The EmptyRateHashingService was always returning false which halted the RatesRetrieverService from going through the full processes.  Since it got halted, it didn't get to the RatesRetriever that actually checks the shipment type for rates support and returns a RateGroup with error and footnote about not supporting rates.  I've updated EmptyRateHashingService.IsRatingField to return true so that the process continues.