TP#12249: Crash when hitting a letter in shipping panel provider dropdown then hitting tab
There's a know issue with machines that have .Net 4.6+ installed where an exception will be thrown by a WPF ComboBox in certain scenarios.  One of which is our provider drop down.  Select None, hit the letter F (for FedEx), the combo box updates, then hit TAB, and it will crash.

Updated the ShippingPanelControl to use the ShipWorksComboBox for the provider drop down, and updated the ShipWorksComboBox to implement the workaround referenced in the link below.

https://connect.microsoft.com/VisualStudio/feedback/details/1660886/system-windows-controls-combobox-coerceisselectionboxhighlighted-bug