Update behavior in dialog that is shown when there a stores in ShipWorks that Tango doesn't know about.
Before, when over the channel limit or stores where out of sync, the dialog would show all channels both tango and shipworks know about. Now, when over the limit, we only show Tango stores. When the stores are out of sync, we show the stores that are in SW but not in tango.