TP#12954: Clicking the "Suggestions" link would act like it's not working.
The Shipment.ShipAddressValidationSuggestionCount was not being reset to 0 when the address was changed and no entries in the ValidatedAddress table.  This would cause the Suggestions like to display, but it had nothing to actually display.  I've updated the ShippingManagerWrapper.SaveShipmentsToDatabase to update these count field based on number of validated addresses.  And if that count is 0, reset the status to NotChecked so that "Validate" will be displayed.