In CustomerLicenseActivationDlgViewModel, change type of OkClickCommand to ICommand.
Create the command only once, instead of every time the property is accessed.

Recommended by Andrew. "This is because all of our test assemblies are .NET 4.5 while our main assemblies are 4.0. MvvmLight (where the RelayCommands live) have different version numbers for .NET 4.0 and 4.5 so it makes referencing the same assemblies really difficult.".