TP#13056: Admin user shouldn't have access to manage profiles in certain scenarios
Updated as per:
If order with no shipment is selected, disable the button.
If a single order is selected for which the user has shipment create/edit/process rights, enable the button.
If a single order is selected for which the user does NOT shipment create/edit/process rights, disable the button.
If multiple orders are selected and the user has permission for each, enable the button.
If multiple orders are selected and the user does NOT have permission for at least one, disable the button.