TP:13064 - Don't crash when open account suggests an address.
When an address was suggested, both ok and cancel acted as if the user clicked cancel and SW would crash.
I fixed this by making the DialogResult of the ok button be OK instead of closing the from in the code behind.
After this was fixed, since UPS wasn't giving us a street address, when the user clikced ok, the street
address was blanked out which resulted in another error. The WSDL says they can give us a street address,
so now I'm only setting street address if it is not null.

At this point, cancel was still causing a crash. I fixed this by raising a UpsOpenAccountException if the user cancels.