TP#13083: Rates never finishing in rating panel
SelectRate was always setting ServiceType no matter if it was a different value.  ServiceType needs to force updating the value so that combo boxes refresh correctly.  This kicks off a get rates call.  When get rates finishes it selects a rate in the rating panel which updates service type...rinse and repeat.  Now in the shipment view model base we only update ServiceType if the selected rate's value is different.

Also updated packaging type to use the first value in the package types list if the previously selected package type doesn't exist in the list.