TP#13083:  Rates never finishing in rating panel (part 2)
When handling the rates retrieved, we needed to hide the rates in the rate panel so we were clearing the Rates list which fired off a SelectedRate(null) that kicked off that whole pipeline.  Now, hiding the grid when loading instead of clearing the rates.  The rates collection gets updated when we received the rates retrieved msg and the grid becomes visible again.