TP#13116: Crash when selecting BestRate in quick ship panel
ShippingAccountRequiredForRatingFootnoteViewModel had a ShippingSettingsDlg constructor parameter that was not registered with AutoFac.  No other code was using a Func to get a shipping settings dlg, so updated this one to not use registration and to be the same as the other uses to be consistent.  If we decide to change them at some later point we can.

BooleanConverter and BooleanToVisibilityConverter needed to be marked as excluded from obfuscation.