TP#12906 Retain username and password for data source
When you are in ODBC control, enter a username and password, select another datasource and come back, the username
and password are retained.

Also, if you enter a username and password and switch to another datasource where you haven't enterred a username
and password, username and password are blanked out.