TP#13360: BestRate got "stuck" when trying to use Account Address, and didn't become unstuck until selecting a different order
The AddressViewModel was attempting to load the account by EntityID, but for BestRate we use the NullCarrierAccount which does not have a valid entity id.  This caused all processing to stop in the view models.  Now checking to see if the person Entity is the NullCarrierAccount as well as for null and not trying to load from the db.