TP#13361: When a customs weight is changed, the rates do not refresh.
The view model content weight property was named ShipmentContentWeight which does not match up to the ShipmentEntity ContentWeight field, so it did not match the rating field names and didn't cause a rates refresh.  Renamed ShipmentContentWeight to ContentWeight.

The view models were also using TotalWeight instead of ContentWeight, which the carrier service controls use.  Updated the view models to bind to and update ContentWeight and removed the TotalWeight property.

Integration tests would deadlock occassionally, so wrapped the Save methods in a SqlAdapterRetry.