TP#13426: Restricted shipment types showing in shipping panel provider dropdown
The provider drop down is populated when ShipWorks is opened.  It is getting the list of shipment types before EditionManager has loaded restrictions.  So, added EditionManager.UpdateRestrictions as the last call of UserSession.InitializeForCurrentSession.

Also added sending EnabledCarriersChangedMessage when Store and Customer licenses are refreshed so that the shipping panel can update it's list of shipment types.

Also updated ShipmentTypeProvider to get it's list of shipment types in a similar way as the ship order dlg.