TP#13677 Remove OdbcOrderElementFactory and handle order element creation differently
Added constructors for OrderItemEntity and OrderItemAttributeEntity that take an order and orderitem respectively.
They then set the appropriate field with the passed in property and set the defaults to null
I moved the note logic out of the factory and into new Note and OrderRepository classes.
Now, newing up the elements can be done in one line and there is no need for a factory.
Update IOdbcOrderDetailLoader to not take the factory