TP#15112: Remove check for interapptive user when registering Amazon filters and enforcers.
We had a check where if you were an interapptive user, we would not register some of the Amazon filters and enforcers which caused our testing to be incorrect. I've removed this so now you can only use magic keys to enable this behavior.
Also, the style of the Same Day footnote was not consistent with the WinForms rating panel, so that was fixed, as well.