TP#13552: Cache IsCustomerLicenseSupported to improve Shipping Dialog load time
When working with large numbers of shipments in the shipping dialog, SqlSchemaVersion.IsCustomerLicenseSupported gets called a large number of times. Each time this is called, the installed schema version is retrieved from the database. Since this value shouldn't change except after an upgrade, we now cache it and reset the cache each time a databse is initialized