TP#15504 When updating non-webreg, CustomerKey was empty. This caused a crash.
When Assembly failes to update and succeeds the second time, VersionSepecificUpdates were not applying because they were selected based on Schema version.
CustomerKey is populated with a VersionSpecificUpdate
Updated VersionSpecificUpdate to use assembly version so the VersionSpecificUpdates actually run.

This doesn't fix the problem 100%. This only addresses the issue when DeployAssembly failed.  If there is a crash during the upgrade after the assembly was deployed and prior to all the VersionSpecificUpdates from running, SW will still miss some version specific updates.

Also, added logging to version specific updates...